+++ admin/manage-users.php
@@ -1,5 +1,7 @@
 <?php include('partials/menu.php'); ?>
-
+    <style>
+        <?php include '../css/admin.scss'; ?>
+    </style>
         <!-- Main Content Section Starts -->
         <div class="main-content">
             <div class="wrapper">
@@ -42,9 +44,20 @@
                         echo $_SESSION['change-pwd'];
                         unset($_SESSION['change-pwd']);
                     }
-
+                echo "<script>
+                    setTimeout(function(){
+                        var element = document.getElementById('tempDiv');
+                        if (element) {
+                            element.style.transition = 'opacity 1s';
+                            element.style.opacity = '0';
+                            setTimeout(function(){
+                                element.parentNode.removeChild(element);
+                            }, 1000);
+                        }
+                    }, 2000);
+                  </script>";
                 ?>
-                <br><br><br>
+                <br>
 
                 <a href="add-user.php" class="btn-primary">Add User</a>
 
@@ -85,7 +98,7 @@
                                         <td>
                                             <a href="<?php echo SITEURL; ?>admin/update-password.php?id=<?php echo $id; ?>" class="btn-primary">Change Password</a>
                                             <a href="<?php echo SITEURL; ?>admin/update-user.php?id=<?php echo $id; ?>" class="btn-secondary">Update User</a>
-                                            <a href="<?php echo SITEURL; ?>admin/delete-user.php?id=<?php echo $id; ?>" class="btn-danger">Delete User</a>
+                                            <a href="<?php echo SITEURL; ?>admin/delete-user.php?id=<?php echo $id; ?>" onclick="return confirm('Are you sure you want to delete this user?')" class="btn-danger">Delete User</a>
                                         </td>
                                     </tr>
